requirements.txt
langchain==1.2.13
langgraph==1.1.3
requests>=2.31.0
python-dotenv==1.2.2
langchain-groq==1.1.2
streamlit==1.55.0
langchain-community==0.4.1
qdrant-client==1.17.1
langchain-qdrant==1.1.0
langchain-huggingface==1.2.1
sentence-transformers==5.3.0
bs4==0.0.2
rank-bm25==0.2.2
fastembed>=0.4.0